rename jsonrpcserver to jsonrpc_server and fix related imports and issues
- Rename jsonrpcserver.py to jsonrpc_server.py for consistent naming
- Replace armor with asyncio.shield for cancellation protection
- Fix get_running_loop to get_event_loop for compatibility
- Fix subscription.py import from old jsonrpcserver package
- Fix indentation in test_jsonrpc.py and method name in stats.py
